feat: 新建插件api應用程式（ext_api）
