+++ apps/ext_api/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CeApiConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "apps.ext_api"
