+++ apps/ext_api/urls.py
@@ -1,22 +1,22 @@
 from django.urls import path
 from . import views
 
-app_name = "rewards"
+app_name = "ext_api"
 
 
 urlpatterns = [
     path(
-        "api/rewards/",
+        "rewards/",
         views.get_rewards,
         name="get_rewards",
     ),
     path(
-        "api/rewards/category/<category>",
+        "rewards/category/<category>",
         views.get_category_rewards,
         name="get_category_rewards",
     ),
     path(
-        "api/rewards/scope/<scope>",
+        "rewards/scope/<scope>",
         views.get_merchant_rewards,
         name="get_merchant_rewards",
     ),
